refactor: refactored createProgressNode logic (#486)
